Fix parsing of durations and works to use doubles

+++ src/Immutable.ProjectModel/FieldKinds.DurationFieldKind.cs
@@ -24,7 +24,7 @@ namespace Immutable.ProjectModel
 
                     text = text.Trim();
 
-                    if (int.TryParse(text, out var actualValue))
+                    if (double.TryParse(text, out var actualValue))
                     {
                         value = TimeSpan.FromDays(actualValue);
                         return true;
